### 2022-08-06 #4657 @editorialbot
[PRE REVIEW]: overviewR - Easily Explore Your Data in R
**Submitting author:** @cosimameyer (<a href="http://orcid.org/0000-0002-7472-2298">Cosima Meyer</a>)
**Repository:** https://github.com/cosimameyer/overviewR
**Branch with paper.md** (empty if default branch): master
**Version:** v0.0.10
