+++ README.md
@@ -14,7 +14,7 @@ badge](https://img.shields.io/badge/overviewR-ready%20to%20use-brightgreen)](htt
 [![R
 badge](https://img.shields.io/badge/Build%20with-♥%20and%20R-blue)](https://github.com/cosimameyer/overviewR)
 <!--[![downloads](https://cranlogs.r-pkg.org/badges/overviewR)](https://www.r-pkg.org/badges/version/overviewR)-->
-[![total_downloads](https://cranlogs.r-pkg.org/badges/grand-total/overviewR)](https://www.r-pkg.org/badges/version/overviewR)
+[![total_downloads](https://cranlogs.r-pkg.org/badges/grand-total/overviewR)](https://cranlogs.r-pkg.org/badges/grand-total/overviewR)
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/overviewR)](https://cran.r-project.org/package=overviewR)
 [![DOI](https://zenodo.org/badge/269955026.svg)](https://zenodo.org/badge/latestdoi/269955026)
 [![license](https://img.shields.io/badge/license-GPL--3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0.en.html)
