### 2022-08-11 3b8733b
Changes to enable changing the `dot_size` and the `fontsize`
