+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   getting-started: getting-started.html
   overviewR_vignette: overviewR_vignette.html
-last_built: 2022-08-11T04:02Z
+last_built: 2022-08-14T05:31Z
 
