### 2022-08-15 73f9403
Update invalid DOIs
