+++ paper/paper.bib
@@ -12,7 +12,7 @@
   pages={1--22},
   year={2021},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.1080/09557571.2021.1980498}
+  doi={10.1080/09557571.2021.1980498}
 }
 
 @article{meyer2022power,
@@ -23,7 +23,7 @@
   number={4},
   year={2022},
   publisher={Oxford University Press},
-  doi={https://doi.org/10.1093/jogss/ogac009}
+  doi={10.1093/jogss/ogac009}
 }
 
 @phdthesis{meyer2021power,
@@ -106,7 +106,6 @@
   year = {2019},
   journal = {{The R Journal}},
   doi = {10.32614/RJ-2019-033},
-  url = {https://doi.org/10.32614/RJ-2019-033},
   pages = {347--369},
   volume = {11},
   number = {2}
