### 2022-08-17 #4657 @samhforbes
Thanks got submitting to JOSS @cosimameyer. I have been in touch with some possible reviewers, and will add them here as and when they accept the invitation. Until then I look forward to going through this process with you!
