### 2022-08-25 #4657 @samhforbes
Hi @jaeyk @pmyteh would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Please note this is a public issue, and your responses will also be public.
