### 2022-09-06 cosimameyer/overviewR#18 @drmowinckels
mention and compare package to other similar packages
Hi!

As part of your submission to JOSS (https://github.com/openjournals/joss-reviews/issues/4740), I'd recommend adding a section to your paper (and maybe even the readme) comparing your package functionality to packages with similar or overlapping functionality, like [skimr](https://cran.r-project.org/web/packages/skimr/vignettes/skimr.html), [DataExplorer](https://cran.r-project.org/web/packages/DataExplorer/vignettes/dataexplorer-intro.html) and [naniar](https://cran.r-project.org/web/packages/naniar/vignettes/getting-started-w-naniar.html). I think this will give new users an idea of what this package might offer them beyond the options they might already have.
