### 2022-09-06 cosimameyer/overviewR#19 @drmowinckels
add community guidelines
as art of your JOSS review (https://github.com/openjournals/joss-reviews/issues/4740), I'd like to see some community guidelines in your readme on

1) Where to report bugs
2) How to contribute to the package
3) How to seek support
