### 2022-09-06 cosimameyer/overviewR#21 @drmowinckels
dont reiterate devtools installs in code sections
As part of (https://github.com/openjournals/joss-reviews/issues/4740)

dont re-iterate the devtools install commands in subsequent readme demo sections. Readme usage sections should assume the tools was installed at the beginning, and not have install bits on top of every code section. It can be confusing for novices to think the tool needs several re-installs.
