### 2022-09-06 cosimameyer/overviewR#22 @drmowinckels
object 'toyota_red' is not created in a visible part of the readme
Readme code fails when trying to plot the 'toyota_red' object. Make sure that the object is created with visible code.

part of https://github.com/openjournals/joss-reviews/issues/4740
