### 2022-09-06 cosimameyer/overviewR#23 @drmowinckels
object 'toydata_with_na' not created in visible code
Readme code fails when trying to plot the 'toyota_red' object. Make sure that the object is created with visible code.

part of https://github.com/openjournals/joss-reviews/issues/4740
