### 2022-09-06 cosimameyer/overviewR#24 @drmowinckels
recommendation: rename _print to _latex
I'd recommend renaming the `_print` functions to `_latex`. This will make it more clear that the functions create latex output.
Also, it creates a clear distinction from R print methods, that other developers and more advanced users might be familiar with, which is quite different.

as part of https://github.com/openjournals/joss-reviews/issues/4740
