### 2022-09-06 f5cd281
Rm devtools from code section (it was a left-over from the development process); we kept `magrittr` because it's relevant in the specific code section (and not generally for the use of overviewR) - but I'm open for suggestions :)
Re #21
