### 2022-09-06 d49612e
Add community guidelines (re #19)
