+++ README.Rmd
@@ -680,4 +680,17 @@ AGO & 1990 - 1992 &  \\
 
 The hex sticker is generated by ourselves using the [```hexSticker```](https://github.com/GuangchuangYu/hexSticker) package.
 
+# How to reach out? 
+
+1. Where do I report bugs?
+
+Simply [open an issue](https://github.com/cosimameyer/overviewR/issues/new) on GitHub.
+
+2. How do I contribute to the package?
+
+If you have an idea (but no code yet), [open an issue](https://github.com/cosimameyer/overviewR/issues/new) on GitHub. If you want to contribute with a specific feature and have the code ready, fork the repository, add your code, and create a pull request.
+
+3. Do you need support?
+
+The easiest way is to [open an issue](https://github.com/cosimameyer/overviewR/issues/new) - this way, your question is also visible to others who may face similar problems.
 
