### 2022-09-06 6bf36ab
Include code for `toydata_with_na` (re #23)
