+++ vignettes/overviewR_vignette.Rmd
@@ -430,7 +430,9 @@ overview_heat(toydata_red,
 
 `overview_na` is a simple function that provides information about the content of all variables in your data, not only the time and scope conditions. It returns a horizontal ggplot bar plot that indicates the amount of missing data (NAs) for each variable (on the y-axis). You can choose whether to display the relative amount of NAs for each variable in percentage (the default) or the total number of NAs. 
 
-```{r, include=FALSE}
+For this purpose, we first artificially reduced our `toydata`.
+
+```{r, include=TRUE}
 toydata_with_na <- toydata %>%
   dplyr::mutate(
     year = ifelse(year < 1992, NA, year),
