### 2022-09-06 f28d0b9
Include code to generate `toydata_red`
