### 2022-09-06 cosimameyer/overviewR#21 @cosimameyer
Thanks for the super helpful suggestion. This was a left over from the development process (when the function was just implemented in the development version). I removed it but kept the call of `magrittr` (because it's not essential for `overviewR` but necessary in this specific code chunk).
The changes are included in PR #25
