### 2022-09-06 67d1b71
Rename function (re #24)
