+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   getting-started: getting-started.html
   overviewR_vignette: overviewR_vignette.html
-last_built: 2022-08-15T07:19Z
+last_built: 2022-09-06T20:07Z
 
