+++ man/overview_latex.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/overview_print.R
-\name{overview_print}
-\alias{overview_print}
-\title{overview_print}
+% Please edit documentation in R/overview_latex.R
+\name{overview_latex}
+\alias{overview_latex}
+\title{overview_latex}
 \usage{
-overview_print(
+overview_latex(
   obj,
   title = "Time and scope of the sample",
   id = "Sample",
@@ -65,7 +65,7 @@ Produces a 'LaTeX' output for output obtained via
 data(toydata)
 
 overview_object <- overview_tab(dat = toydata, id = ccode, time = year)
-overview_print(
+overview_latex(
   obj = overview_object,
   title = "Some nice title",
   crosstab = FALSE
@@ -80,7 +80,7 @@ overview_ct_object <- overview_crosstab(
   id = ccode,
   time = year
 )
-overview_print(
+overview_latex(
   obj = overview_ct_object,
   title = "Some nice title for a cross tab",
   crosstab = TRUE,
