+++ R/overview_latex.R
@@ -1,4 +1,4 @@
-#' @title overview_print
+#' @title overview_latex
 #'
 #' @description Produces a 'LaTeX' output for output obtained via
 #'    \code{overview_tab} and \code{overview_crosstab}
@@ -30,7 +30,7 @@
 #' data(toydata)
 #'
 #' overview_object <- overview_tab(dat = toydata, id = ccode, time = year)
-#' overview_print(
+#' overview_latex(
 #'   obj = overview_object,
 #'   title = "Some nice title",
 #'   crosstab = FALSE
@@ -45,15 +45,14 @@
 #'   id = ccode,
 #'   time = year
 #' )
-#' overview_print(
+#' overview_latex(
 #'   obj = overview_ct_object,
 #'   title = "Some nice title for a cross tab",
 #'   crosstab = TRUE,
 #'   cond1 = "Name of first condition",
 #'   cond2 = "Name of second condition"
 #' )
-#' @export overview_print overview_latex
-#' @alias overview_print
+#' @export overview_latex overview_latex
 #' @importFrom dplyr "%>%"
 
 
@@ -197,4 +196,4 @@ overview_latex <-
     }
   }
 
-overview_print <- overview_latex
+overview_latex <- overview_latex
