+++ R/overview_crosstab.R
@@ -13,7 +13,7 @@
 #' @param id Scope (e.g., country codes or individual IDs)
 #' @param time Time (e.g., time periods given by years, months, ...)
 #' @return A data frame object that contains a summary of the data set that can
-#'       later be converted to a 'LaTeX' output using \code{overview_print}
+#'       later be converted to a 'LaTeX' output using \code{overview_latex}
 #' @examples
 #' data(toydata)
 #' overview_crosstab(
