### 2022-09-06 cosimameyer/overviewR#24 @cosimameyer
Thanks so much for this suggestion - that's excellent! I didn't think about it this way (we first envisioned the `overview_print` function to become a much broader function but it remained an `overview_latex` function as it is now and you are right - this name serves its purpose much better 😊). We discussed it internally and are very happy to go forward with your suggestion in PR #26
