+++ README.md
@@ -13,17 +13,15 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 badge](https://img.shields.io/badge/overviewR-ready%20to%20use-brightgreen)](https://github.com/cosimameyer/overviewR)
 [![R
 badge](https://img.shields.io/badge/Build%20with-♥%20and%20R-blue)](https://github.com/cosimameyer/overviewR)
-<!--[![downloads](https://cranlogs.r-pkg.org/badges/overviewR)](https://www.r-pkg.org/badges/version/overviewR)-->
-[![total_downloads](https://cranlogs.r-pkg.org/badges/grand-total/overviewR)](https://cranlogs.r-pkg.org/badges/grand-total/overviewR)
-[![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/overviewR)](https://cran.r-project.org/package=overviewR)
-[![DOI](https://zenodo.org/badge/269955026.svg)](https://zenodo.org/badge/latestdoi/269955026)
+[![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/overviewR)](https://cran.r-project.org/package=overviewR)
 [![license](https://img.shields.io/badge/license-GPL--3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0.en.html)
+[![metacran
+downloads](https://cranlogs.r-pkg.org/badges/overviewR)](https://cran.r-project.org/package=overviewR)
 <!-- [![Rdoc](https://www.rdocumentation.org/badges/version/overviewR)](https://www.rdocumentation.org/packages/overviewR) -->
-<!-- [![metacran downloads](https://cranlogs.r-pkg.org/badges/overviewR)](https://cran.r-project.org/package=overviewR) -->
 <!-- [![cran checks](https://cranchecks.info/badges/summary/overviewR)](https://cran.r-project.org/web/checks/check_results_overviewR.html) -->
 <!-- [![](https://cranlogs.r-pkg.org/badges/version/overviewR)](https://www.r-pkg.org/badges/version/overviewR) -->
 <!-- [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0) -->
-<!-- [![Last-changedate](https://img.shields.io/badge/last%20change-2021--06--04-green.svg)](/commits/master) -->
+<!-- [![Last-changedate](https://img.shields.io/badge/last%20change-2022--09--06-green.svg)](/commits/master) -->
 <!-- badges: end -->
 
 [**You can access the CheatSheet for overviewR
@@ -49,7 +47,8 @@ the following functions:
 -   `overview_crossplot` is an alternative to visualize a cross table (a
     way to present results from `overview_crosstab`)
 -   `overview_heat` plots a heat map of your time line
--   `overview_na` plots an overview of missing values by variable
+-   `overview_na` plots an overview of missing values by variable (both
+    by row and by column)
 -   `overview_overlap` plots comparison plots (bar graph and Venn
     diagram) to compare to data frames
 
@@ -93,13 +92,13 @@ cross-sectional format on 5 countries, covering the period 1990-1999.
 ``` r
 data(toydata)
 head(toydata)
-#>   ccode year month      gdp population
-#> 1   RWA 1990   Jan 24180.77  14969.988
-#> 2   RWA 1990   Feb 23650.53  11791.464
-#> 3   RWA 1990   Mar 21860.14  30047.979
-#> 4   RWA 1990   Apr 20801.06  19853.556
-#> 5   RWA 1990   May 18702.84   5148.118
[…]
