+++ README.Rmd
@@ -37,8 +37,8 @@ lapply(pkgs, library, character.only = TRUE)
 [![R badge](https://img.shields.io/badge/Build%20with-♥%20and%20R-blue)](https://github.com/cosimameyer/overviewR)
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/overviewR)](https://cran.r-project.org/package=overviewR)
 [![license](https://img.shields.io/badge/license-GPL--3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0.en.html)
+[![metacran downloads](https://cranlogs.r-pkg.org/badges/overviewR)](https://cran.r-project.org/package=overviewR)
 <!-- [![Rdoc](https://www.rdocumentation.org/badges/version/overviewR)](https://www.rdocumentation.org/packages/overviewR) -->
-<!-- [![metacran downloads](https://cranlogs.r-pkg.org/badges/overviewR)](https://cran.r-project.org/package=overviewR) -->
 <!-- [![cran checks](https://cranchecks.info/badges/summary/overviewR)](https://cran.r-project.org/web/checks/check_results_overviewR.html) -->
 <!-- [![](https://cranlogs.r-pkg.org/badges/version/overviewR)](https://www.r-pkg.org/badges/version/overviewR) -->
 <!-- [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0) -->
@@ -456,7 +456,7 @@ This function allows to compare two data sets. We are currently working on an ex
 
 At the current development stage, the function works as follows:
 
-```{r, eval=FALSE}
+```{r, eval=TRUE, out.width = '50%', fig.align='center'}
 library(dplyr)
 
 # Subset one data set for comparison
@@ -472,12 +472,13 @@ overview_overlap(
 ```
 
 
-<p align="center">
+<!--<p align="center">
 <img src='man/figures/bar.png' height="400"/>
-</p>
+</p>-->
 
 Or a Venn diagram
-```{r, eval = TRUE}
+
+```{r, eval = TRUE,out.width = '50%', fig.align='center'}
 overview_overlap(
   dat1 = toydata,
   dat2 = toydata2,
@@ -488,9 +489,11 @@ overview_overlap(
 ```
 
 
-<p align="center">
+<!--<p align="center">
 <img src='man/figures/venn.png' height="400"/>
 </p>
+-->
+
 
 ## Compatibilities with other packages 
 
@@ -680,21 +683,22 @@ AGO & 1990 - 1992 &  \\
 <img src='man/figures/extension1.png' height="150"/>
 </p>
 
-# Credits
-
-The hex sticker is generated by ourselves using the [```hexSticker```](https://github.com/GuangchuangYu/hexSticker) package.
 
 # How to reach out? 
 
-1. Where do I report bugs?
+### Where do I report bugs?
 
 Simply [open an issue](https://github.com/cosimameyer/overviewR/issues/new) on GitHub.
 
-2. How do I contribute to the package?
+### How do I contribute to the package?
 
 If you have an idea (but no code yet), [open an issue](https://github.com/cosimameyer/overviewR/issues/new) on GitHub. If you want to contribute with a specific feature and have the code ready, fork the repository, add your code, and create a pull request.
 
-3. Do you need support?
+### Do you need support?
 
 The easiest way is to [open an issue](https://github.com/cosimameyer/overviewR/issues/new) - this way, your question is also visible to others who may face similar problems.
+
+# Credits
+
+The hex sticker is generated by ourselves using the [```hexSticker```](https://github.com/GuangchuangYu/hexSticker) package.
 
