+++ dev/pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 2.7.3
-pkgdown: 1.6.1
+pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   getting-started: getting-started.html
   overviewR_vignette: overviewR_vignette.html
-last_built: 2021-08-13T08:06Z
+last_built: 2022-09-06T20:42Z
 
