### 2022-09-07 d3239af
Update README with a table (suggested in #18)
