### 2022-09-07 fdf6024
Update paper with overview and contrasting other packages as suggested in #18
