+++ paper/paper.md
@@ -32,7 +32,7 @@ The main advantages of overviewR are threefold: 1) it allows users to quickly un
 
 # Statement of need
 
-overviewR has already been used in the data wrangling process as well as to generate an illustrative overview of the sample in various academic publications and research [@hammerschmidt2021moving; @hammerschmidt2021foreign; @meyer2022power; @meyer2021power]. While not limited to these applications, we see its main advantage for researchers and practitioners working with time-series cross-sectional data and its power to illustratively visualize key information of the data. When comparing with other exploratory data analysis packages available (for instance DataExplorer [@dataexplorer2020], gtsummary [@gtsummary2022], dlookR [@ryu2021dlookr], Hmisc [@hmisc2022], naniar [@naniar2021], smartEDA [@putatunda2019smarteda], skimr [@skimr2022], or summarytools [@comtois2022package]), we do not see the key functions of overviewR fully covered. A detailed comparison which contrasts the core functionalities and advantages of overviewR is outlined in  Figure \ref{overview} below.
+overviewR has already been used in the data wrangling process as well as to generate an illustrative overview of the sample in various academic publications and research [@hammerschmidt2021moving; @hammerschmidt2021foreign; @meyer2022power; @meyer2021power]. While not limited to these applications, we see its main advantage for researchers and practitioners working with time-series cross-sectional data and its power to illustratively visualize key information of the data. When comparing with other exploratory data analysis packages available (for instance DataExplorer [@dataexplorer2020], gtsummary [@gtsummary2022], dlookR [@ryu2021dlookr], Hmisc [@hmisc2022], naniar [@naniar2021], smartEDA [@putatunda2019smarteda], skimr [@skimr2022], or summarytools [@comtois2022package]), we do not see the key functions of overviewR fully covered. A detailed comparison that contrasts the core functionalities and advantages of overviewR is outlined in  Figure \ref{overview} below.
 
 ![Comparison of overviewR with available R packages\label{overview}](figures/overview_packages.png)
 
