### 2022-09-07 cosimameyer/overviewR#18 @cosimameyer
Thanks so much for the suggestion! I added an overview in the paper (as well as in the README) in PR #28
