### 2022-09-10 #4740 @milenamilena
I have completed my review, and I found this package very useful and holistic in giving an overview of data. The functions are very clear and well explained. The latex tables were beneficial, as the format is very clear. I also found very interesting the use of the Venn diagram and the ability to customise some gglots. I agree with the comments made by @drmowinckels and am happy to see that they were addressed. It is very promising to see that there are open issues in the repository, mainly for further enhancement. The whole paper is very explanatory, even for a naive R user.
