### 2022-09-12 cosimameyer/overviewR#29 @drmowinckels
why have separate path and file arguments to the _latex function?
This is part of the https://github.com/openjournals/joss-reviews/issues/4740 .

I was wondering why there is a need to specify both destination folder and filename for saving a file? This is quite uncommon when saving files in R, where we would usually provide the entire path as a single argument.
