### 2022-09-12 cosimameyer/overviewR#29 @cosimameyer
Thanks for the suggestion, @drmowinckels :)

We created the function to work with one base path for multiple outputs. Coming from a social sciences background, you might want to generate overview tables for multiple subsets of your data set. Creating a file path upfront in the script and adding the `file_path` object in the `overview_latex` function (thanks so much for suggesting the new name, I love it! 🥰 ) allows us to avoid path copying.

```
file_path <- 'some/file_path/'

overview_latex(
   obj = overview_table1,
   path = file_path,
   file = 'plot1.tex'
)

overview_latex(
   obj = overview_table2,
   path = file_path,
   file = 'plot2.tex'
)
```

An alternative that I can think of, on the top of my head, would be to rewrite the function as follows:

```
file_path <- 'some/file_path/'

overview_latex(
   obj = overview_table1,
   path = paste0(file_path, 'plot1.tex')
)

overview_latex(
   obj = overview_table2,
   path = paste0(file_path, 'plot2.tex')
)
```

I agree that it makes it shorter, but I'm not sure if it makes it cleaner. But I'm curious what you think 😊
