### 2022-09-12 cosimameyer/overviewR#29 @drmowinckels
I think there are pros and cons for each, and having created a package that does indeed take a folder as a destination I understand what you are thinking. I too am from social science (psychology), so I totally get where you are coming from.

At the same time, I think a folder makes sense _if_ the function it self creates multiple files. But if a single call to the function creates a single file, then giving the full (or relative) path makes more sense. And then its up to the user to do as you do.

Think of how other writing functions work, where they take a single path for a single file. write.table, readr::write_tsv, etc. We are used to having to provide a single path for a single function call already. As a user, if I were just writing a single file, I'd find it strange to provide both a path and  destination folder, I'd for sure do it incorrectly at the first try because of this. From an R point of view, I think a single path makes sense.

That being said. I am not going to force a change on this at all, I merely wanted a discussion around it, as I found it odd. It is not a big deal at all, and if you are happy with it, then go for it. If you, after reading this comment, still want to keep it as is, then you can close the comment, and I'll finish my review.
