### 2022-09-12 cosimameyer/overviewR#29 @cosimameyer
Thanks for your response and for sharing your thoughts. We find the reasoning very convincing and highly appreciate getting an outside perspective on our package's functionality :)
We changed the argument from `file` and `path` to a single argument `file_path` in PR #30 :)
