+++ dev/pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   getting-started: getting-started.html
   overviewR_vignette: overviewR_vignette.html
-last_built: 2022-09-07T17:50Z
+last_built: 2022-09-12T20:15Z
 
