### 2022-09-13 #4740 @drmowinckels
I have completed my review and am happy with the responses and changes made to the package and paper. The package functionality is well documented and covers something that has no real alternative in the R eco-system as I know it. Its clearly well-thought out and tested, and the implementation is great.

@samhforbes I believe both reviewers are happy with the package and paper status.
