### 2022-09-13 #4740 @samhforbes
Hi @milenamilena @drmowinckels thanks for your hard work and suggestions. We appreciate both your thoroughness and the time spent reviewing and improving packages for JOSS.

@cosimameyer it looks like things are all coming together!

There's a couple of things we need to do before ticking off the final box. So can you please:
- Let me know the final version number (I think you are now on 0.0.11, don't know if you are planning to go to 12 before this is finished)
- Can you create a tagged release and archive a JOSS accepted version somewhere with a DOI, and paste the DOI here

Thanks!
