### 2022-09-13 #4740 @cosimameyer
Thanks for the update and for putting your time into it - it’s highly appreciated!
This review process is one of the best I've experienced (not just because of the outcome 😉) I think the discussions here and in the issues as well as the general feedback you get are so valuable! So thanks to all you for volunteering at the journal and for sharing your expertise ☺️

@samhforbes here's the missing information:
- We upgraded overviewR to 0.0.12 (can you upgrade it here? I'm not sure if I can do it myself)
- I archived the GitHub repo at Zenodo, the DOI is 10.5281/zenodo.7075665
