### 2022-09-13 #4740 @cosimameyer
@oliviaguest  Thanks for following up! I'm not sure if it really is a missing DOI (I didn't change anything in the references). The first two were already part of our initial submission (and, as far as I know, they don't have a DOI) and the third one (`naniar`) was added as part of the review process - but I checked the references after making the changes here (https://github.com/openjournals/joss-reviews/issues/4740#issuecomment-1239705390) and the editorial bot did not complain. Let me know what you think 😊
