### 2022-09-14 #4740 @cosimameyer
As far as I understand the log (and I might be completely off here), there is an invalid JATS. Is it the DOI that I provided that causes the problems? I'm not sure but I believe it [should be fine](https://zenodo.org/record/7075665). If there's anything that I can/should do, let me know 😊
