### 2022-09-20 #4740 @xuanxu
Footnotes are not allowed in the Institutions list. The error is caused by the syntax used here: https://github.com/cosimameyer/overviewR/blob/master/paper/paper.md?plain=1#L16
