### 2022-09-20 #4740 @cosimameyer
Thanks for the update! I'll change it - is there a way to include the footnote info somewhere? Or should we leave it out?
