### 2022-09-20 #4740 @xuanxu
> is there a way to include the footnote info somewhere?

Maybe adding it as a second affiliation for both authors?
