+++ .github/workflows/pkgdown.yaml
@@ -13,11 +13,8 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
       - uses: actions/checkout@v2
-
       - uses: r-lib/actions/setup-r@v1
-
       - uses: r-lib/actions/setup-pandoc@v1
-
       - name: Query dependencies
         run: |
           install.packages('remotes')
