### 2022-09-20 #4740 @cosimameyer
Excellent suggestion, thanks @xuanxu.
I archived the new version at @samhforbes, the new DOI is 10.5281/zenodo.7097560 (https://doi.org/10.5281/zenodo.7097560). Can you update it?
