### 2022-09-20 #4740 @kyleniemeyer
Hi @cosimameyer, I just made a PR with a few minor edits to the paper, could you merge that? https://github.com/cosimameyer/overviewR/pull/34

Then I can proceed with accepting the paper.
